requirements.txt
asyncpg==0.20.*
ipython==7.11.*
requests==2.23.*
discord.py==1.3.*
